requirements.txt
Django==4.1.1
docker==2.7.0
requests==2.28.1
psutil==5.9.2
django-simple-history==3.1.1
redis==4.3.4
django-celery-beat==2.1.0
celery==5.2.7